Fixed behaviour of opponent canvas with paused players.
When an opponent being viewed is paused, the viewing player will now see
a 'paused' message overlayed the last sent state. If a viewing player
has joined the game after the viewed player paused (and thus has never
received a game state) an empty board is displayed with 'paused'
overlayed.